feat: load NTPC labor violations and disaster map data
- load_labor_violations_ntpc.py: 14155 勞基法 + 47 性平法 + 4148 職安法 = 18350 rows
- load_labor_disasters.py: 150 TPE GPS disasters + 206 NTPC district disasters
- labor_safety_create_tables.sql: fix lng/lat to DOUBLE PRECISION

+++ scripts/labor_safety_create_tables.sql
@@ -49,8 +49,8 @@ CREATE TABLE labor_disasters_tpe (
     disaster_type   VARCHAR(100),
     deaths          INTEGER DEFAULT 0,
     injuries        INTEGER DEFAULT 0,
-    lng             NUMERIC(11,7),
-    lat             NUMERIC(11,7)
+    lng             DOUBLE PRECISION,
+    lat             DOUBLE PRECISION
 );
 
 -- ──────────────────────────────────────────────
